+++ docs/changes.md
@@ -7,7 +7,7 @@ title: Foundation Change Log
 
 ### v2.1.9@dev {#v2-1-9}
 
-* Allow `Orchestra\Foundation\Application::group()` to mimick `Route::group()` functionality.
+* Allow `Orchestra\Foundation\Application::group()` to mimic `Route::group()` functionality.
 * Allow default routes to be configurable.
 
 ### v2.1.8 {#v2-1-8}